+++ bin/modules/UlsInputCli.py
@@ -102,6 +102,7 @@ class UlsInputCli:
         else:
             aka_log.log.critical(
                 f"{self.name} - Feed ({feed}) not available - Available: {product_feeds}")
+            self.stopEvent.set()
             sys.exit(1)
         return feed
 
@@ -115,6 +116,7 @@ class UlsInputCli:
         else:
             aka_log.log.critical(
                 f"{self.name} - FORMAT ({cliformat}) not available")
+            self.stopEvent.set()
             sys.exit(1)
         return cliformat
 
@@ -492,6 +494,7 @@ class UlsInputCli:
             else:
                 aka_log.log.critical(f" {self.name} - No valid product selected "
                                      f"(--input={self.product}).")
+                self.stopEvent.set()
                 sys.exit(1)
 
             try:
@@ -555,7 +558,8 @@ class UlsInputCli:
             if self.running is False and self.rerun_counter > self.rerun_retries:
                 aka_log.log.critical(f'{self.name} - Not able to start the CLI for '
                                      f'{self.product}. See above errors. '
-                                     f'Giving up after {self.rerun_counter - 2} retries.')
+                                     f'Giving up after {self.rerun_counter - 2} retries. Exiting NOW')
+                self.stopEvent.set()
                 sys.exit(1)
 
     def check_proc(self):
@@ -585,7 +589,7 @@ class UlsInputCli:
         except Exception as my_error:
             if self.run_once:
                 aka_log.log.critical(f"{self.name} - '--endtime' was specified - so stopping now")
-
+                self.stopEvent.set()
                 sys.exit(0)
             else:
                 aka_log.log.error(f'{self.name} - Soemthing really '