+++ docs/CHANGELOG.md
@@ -1,4 +1,5 @@
 # Version History
+
 ## v1.9.1
 |        |                     |
 |--------|---------------------|
@@ -7,11 +8,14 @@
 | Author | mschiess@akamai.com |
 
 - **Minor improvements** 
-  - [docker] bumped Python container version to 3.13.5
+  - Show "module version information" in ULS version output (only when --loglevel is set to DEBUG)
+  - [docker] bumped Python container version to 3.13.6
+  - [docker] bumped Python container version from debian bookworm to trixie
   - [docker] bumped CLI-EAA to branch EME-1167 (new fields added)
   - [docker] bumped CLI-LINODE to branch v0.0.7 (more detailed utilization support)
   
 - **Bugfixes**
+  - Fixed a bug in the INPUTCLI, that could have caused a stale behavior of ULS in rare circumstances. We're exiting now properly.
   - Fixed a bug in the Inputhandling CLI `AttributeError: module 'uls_config.global_config' has no attribute 'edgerc_openap'`
   
 ---